fixed locking that failed fixed cache minor logging fixes

+++ core/cache_response.go
@@ -3,39 +3,38 @@ package core
 import (
 	"context"
 	"errors"
-	
-	"github.com/golang/protobuf/ptypes"
-	"github.com/golang/protobuf/ptypes/timestamp"
+	"time"
+
 	"github.com/hashicorp/go-hclog"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
-	
+
 	pb_core "git.liero.se/opentelco/go-swpx/proto/go/core"
 	"git.liero.se/opentelco/go-swpx/shared"
 )
 
 type ResponseCache interface {
-	Pop(ctx context.Context, hostname, port string, rt pb_core.Request_Type) (*CachedResponse, error)
-	Upsert(ctx context.Context, hostname, port string, rt pb_core.Request_Type, response *pb_core.Response) error
-	Clear(ctx context.Context, hostname, port string, rt pb_core.Request_Type) error
+	Pop(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type) (*CachedResponse, error)
+	Upsert(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type, response *pb_core.Response) error
+	Clear(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type) error
 }
 
 type CachedResponse struct {
-	Hostname    string                                 `bson:"hostname"`
-	Port        string                                 `bson:"port"`
-	RequestType pb_core.Request_Type                   `bson:"request_type"`
-	Response    *pb_core.Response                      `bson:"response,omitempty"`
-	Timestamp   *timestamp.Timestamp                   `bson:"timestamp" json:"timestamp"`
+	Hostname    string            `bson:"hostname"`
+	Port        string            `bson:"port"`
+	AccessId    string            `bson:"access_id"`
+	RequestType string            `bson:"request_type"`
+	Response    *pb_core.Response `bson:"response,omitempty"`
+	Timestamp   time.Time         `bson:"timestamp" json:"timestamp"`
 }
-
 
 func newResponseCache(client *mongo.Client, logger hclog.Logger, conf shared.ConfigMongo) (ResponseCache, error) {
 	col := client.Database(conf.Database).Collection(collectionResponseCache)
 	// Create the model
 	model := mongo.IndexModel{
-		Keys:    bson.M{"hostname": -1, "port": -1},
-		Options: options.Index().SetUnique(true),
+		Keys:    bson.M{"hostname": -1},
+		Options: options.Index().SetUnique(false),
 	}
 	c := &respCacheImpl{
 		client: client,
@@ -46,7 +45,7 @@ func newResponseCache(client *mongo.Client, logger hclog.Logger, conf shared.Con
 	if err != nil {
 		return nil, err
 	}
-	
+
 	return c, nil
 }
 
@@ -56,14 +55,32 @@ type respCacheImpl struct {
 	logger hclog.Logger
 }
 
-
-
-func (c *respCacheImpl) Pop(ctx context.Context, hostname, port string, rt pb_core.Request_Type) (*CachedResponse, error) {
-	res := c.col.FindOne(context.Background(), bson.M{"hostname": hostname, "port": port, "request_type": rt})
+func (c *respCacheImpl) Pop(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type) (*CachedResponse, error) {
+	var filter bson.M
+	if accessId != "" {
+		c.logger.Debug("access id is set, filter by that", "access_id", accessId, "rt", rt.String())
+		filter = bson.M{
+			"request_type": rt.String(),
+			"access_id":    accessId,
+		}
+	} else {
+		c.logger.Debug("access id is not set, filter by that", "access_id", accessId, "rt", rt.String())
+		filter = bson.M{
+			"hostname":     hostname,
+			"port":         port,
+			"request_type": rt.String(),
+		}
+	}
+	opts := &options.FindOneOptions{
+		Sort: bson.M{"_id": -1},
+	}
+	res := c.col.FindOne(context.Background(), &filter, opts)
 	obj := &CachedResponse{}
 
 	if err := res.Decode(obj); err != nil {
+
 		if errors.Is(err, mongo.ErrNoDocuments) {
+			c.logger.Warn("err, not found")
 			return nil, nil
 		}
 
@@ -73,24 +90,27 @@ func (c *respCacheImpl) Pop(ctx context.Context, hostname, port string, rt pb_co
 	return obj, nil
 }
 
-func (c *respCacheImpl) Upsert(ctx context.Context, hostname, port string, rt pb_core.Request_Type, response *pb_core.Response) error {
+func (c *respCacheImpl) Upsert(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type, response *pb_core.Response) error {
+
+	c.logger.Debug("insert into cache", "hostname", hostname, "port", port, "type", rt.String(), "access_id", accessId)
 	_, err := c.col.InsertOne(context.Background(), &CachedResponse{
-		hostname,
-		port,
-		rt,
-		response,
-		ptypes.TimestampNow(),
+		Hostname:    hostname,
+		Port:        port,
+		AccessId:    accessId,
+		RequestType: rt.String(),
+		Response:    response,
+		Timestamp:   time.Now(),
 	})
 
 	if err != nil {
-		c.logger.Error("error saving response in cache","error",  err)
+		c.logger.Error("error saving response in cache", "error", err)
 		return err
 	}
 
 	return nil
 }
 
-func (c *respCacheImpl) Clear(ctx context.Context, hostname, port string, rt pb_core.Request_Type) error {
-	_, err := c.col.DeleteMany(context.Background(), bson.M{"hostname": hostname, "port": port, "request_type": rt})
+func (c *respCacheImpl) Clear(ctx context.Context, hostname, port, accessId string, rt pb_core.Request_Type) error {
+	_, err := c.col.DeleteMany(context.Background(), bson.M{"hostname": hostname, "port": port, "request_type": rt.String(), "access_id": accessId})
 	return err
 }
